batch-analysis/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.4
matplotlib==3.8.2
seaborn==0.13.0
wordcloud==1.9.3
numpy==1.26.2
matplotlib-venn==0.11.9
